requirements.txt
asgiref==3.5.1
Django==4.0.4
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1
